requirements.txt
disnake
asyncio
requests
requests_async
beautifulsoup4
jsonschema
newspaper3k
requests
tldextract
urlextract
validators
lxml
wavelink
humanize
ffmpeg-python
imagehash
art
beanie
async-google-trans-new
pycountry
